Enhancements
I added $DeploymentShare as a parameter, so the root of the MDT share can be defined there and used in the $BackupFolderPath, $ExportPath, and $FileToEdit variables. I also changed the $BackupFileOn to $BackupFile, while also making that a switch instead of a boolean. I edited in PowerShell Studio, and it automatically reformatted the layout. The script still works great, and it successfully reordered my MDT Drivers.